Initial rough implementation for interface for AS
Initial rough implementation for AutoStrategizer (AS) interface. This commit marks the initial steps towards building the AS interface, providing a basic framework to build upon. As we delve deeper into designing and exploring the best interface for AutoStrategizer, anticipate frequent refactoring and enhancements to accommodate evolving requirements
The present version is meant for experimental purposes, acting as a canvas for us to fine-tune the functionality and discover the most effective integration to the library.